Extract utils related to date-time in a separate file datetime_utils.dart [rokwire/illinois-app#5859]
