+++ lib/model/inbox.dart
@@ -1,4 +1,5 @@
 import 'package:collection/collection.dart';
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class InboxMessage {
