+++ lib/service/app_datetime.dart
@@ -18,6 +18,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/service.dart';
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:intl/intl.dart';
 import 'package:timezone/timezone.dart' as timezone;
